You_Summarize/requirements.txt
youtube_transcript_api
google-generativeai
streamlit
python-dotenv
pathlib
googletrans==4.0.0-rc1